+++ app/src/main/java/com/brandonjamesyoung/levelup/data/Quest.kt
@@ -8,9 +8,9 @@ import java.time.Instant
 
 @Entity
 data class Quest(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) var id: Int = 0,
     @ColumnInfo val name: String? = null,
     @ColumnInfo val difficulty: Difficulty = Difficulty.EASY,
     @ColumnInfo val iconName: String? = null,
-    @ColumnInfo val dateCreated: Instant? = null,
+    @ColumnInfo var dateCreated: Instant? = null,
 )