+++ app/src/main/java/com/brandonjamesyoung/levelup/shared/Constants.kt
@@ -23,10 +23,6 @@ const val INIT_HARD_POINTS = 6
 const val INIT_EXPERT_EXP = 1500
 const val INIT_EXPERT_POINTS = 12
 
-enum class FontSize {
-    SMALL, MEDIUM, LARGE
-}
-
 enum class Mode {
-    DEFAULT, SELECT
+    DEFAULT, SELECT, EDIT
 }