+++ app/src/main/java/com/brandonjamesyoung/levelup/viewmodels/NewQuestViewModel.kt
@@ -1,8 +1,7 @@
 package com.brandonjamesyoung.levelup.viewmodels
 
 import android.util.Log
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
+import androidx.lifecycle.*
 import com.brandonjamesyoung.levelup.data.*
 import com.brandonjamesyoung.levelup.di.IoDispatcher
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -17,8 +16,17 @@ class NewQuestViewModel @Inject constructor(
     @IoDispatcher private val ioDispatcher: CoroutineDispatcher,
     private val questRepository: QuestRepository
 ) : ViewModel() {
-    private fun logQuestCreation(quest: Quest) {
-        var logMessage = "Add new ${quest.difficulty} quest with "
+    fun getQuest(id: Int): LiveData<Quest> {
+        return questRepository.observe(id).asLiveData()
+    }
+
+    private fun logQuestSave(quest: Quest, isEdit: Boolean = false) {
+        var logMessage = if (!isEdit) {
+            "Add new ${quest.difficulty} quest with "
+        } else {
+            "Edit quest to ${quest.difficulty} difficulty with "
+        }
+
         logMessage += if (quest.name != null) "name '${quest.name}'" else "no name"
         logMessage += " and "
 
@@ -33,6 +41,17 @@ class NewQuestViewModel @Inject constructor(
 
     fun insert(quest: Quest) = viewModelScope.launch(ioDispatcher) {
         questRepository.insert(quest)
-        logQuestCreation(quest)
+        logQuestSave(quest)
+    }
+
+    fun update(quest: Quest) = viewModelScope.launch(ioDispatcher) {
+        val currentQuest = questRepository.get(quest.id)
+
+        if (quest.dateCreated == null) {
+            quest.dateCreated = currentQuest.dateCreated
+        }
+
+        questRepository.update(quest)
+        logQuestSave(quest = quest, isEdit = true)
     }
 }